requirements.txt
psycopg2==2.9.9
Pillow==9.2.0
python-dotenv==1.0.0
python-telegram-bot==20.0
